+++ etc/triton-bash-completion-types.sh
@@ -1,0 +1,8 @@
+# Functions for Bash completion of some 'triton' option/arg types.
+function complete_tritonprofile {
+    local word="$1"
+    local candidates
+    candidates=$(ls -1 ~/.triton/profiles.d/*.json 2>/dev/null \
+        | sed -E 's/^.*\/([^\/]+)\.json$/\1/')
+    compgen $compgen_opts -W "$candidates" -- "$word"
+}